Continuing work on main UI
